refactor: rename ParkingLot project to ParkingLot Service
