+++ ParkingLot.sln
@@ -1,11 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30413.136
+# Visual Studio Version 17
+VisualStudioVersion = 17.3.32929.385
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ParkingLot", "ParkingLot\ParkingLot.csproj", "{EF35A5A3-8388-41A4-B368-BD763BA56C94}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ParkingLotService", "ParkingLotService\ParkingLotService.csproj", "{EF35A5A3-8388-41A4-B368-BD763BA56C94}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ParkingLotTest", "ParkingLotTest\ParkingLotTest.csproj", "{A8ADF2B5-445D-4A86-8DBF-0350586E4EFC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ParkingLotServiceTest", "ParkingLotServiceTest\ParkingLotServiceTest.csproj", "{A8ADF2B5-445D-4A86-8DBF-0350586E4EFC}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{A5567EEF-9878-4039-BBA7-1A9913F6D2F9}"
 	ProjectSection(SolutionItems) = preProject
