--- ParkingLot/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace ParkingLot
-{
-    using System;
-    public class Class1
-    {
-    }
-}
