+++ ParkingLotService/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace ParkingLotService
+{
+    public class Class1
+    {
+    }
+}
