+++ ParkingLotService/stylecop.ruleset
